Twillio Code/requirements.txt
Flask==2.3.3
requests==2.31.0
twilio==8.5.0
python-dotenv==1.0.0
